split magic to own file

+++ _s0il_/picoc/cstdlib/math.c
@@ -123,7 +123,7 @@ void MathRound(struct ParseState *Parser, struct Value *ReturnValue, struct Valu
 {
     /* this awkward definition of "round()" due to it being inconsistently
      * declared in math.h */
-    ReturnValue->Val->FP = ceil(Param[0]->Val->FP - 0.5);
+    ReturnValue->Val->FP = ceil(Param[0]->Val->FP - (double)0.5);
 }
 
 void MathCeil(struct ParseState *Parser, struct Value *ReturnValue, struct Value **Param, int NumArgs)
